SheetSync-Crypto-Reporter/requirements.txt
gspread==6.1.2
google-auth==2.29.0
requests==2.31.0
python-dotenv==1.0.1
schedule==1.2.1